Cambiado estilo grafico y añadidas animaciones
Aun es necessario añadir funcionalidad a la animacion de golpe

+++ 02-Bubble/TileMap.h
@@ -36,11 +36,11 @@ public:
 
 	bool collisionMoveLeft(const glm::ivec2 &pos, const glm::ivec2 &size) const;
 	bool collisionMoveRight(const glm::ivec2 &pos, const glm::ivec2 &size) const;
-	bool collisionMoveDown(const glm::ivec2 &pos, const glm::ivec2 &size, int *posY) const;
-	bool collisionMoveUp(const glm::ivec2& pos, const glm::ivec2& size, int* posY) const;
+	bool collisionMoveDown(const glm::ivec2 &pos, const glm::ivec2 &size, float *posY) const;
+	bool collisionMoveUp(const glm::ivec2& pos, const glm::ivec2& size, float* posY) const;
 
-	bool canClimbUp(const glm::ivec2& pos, const glm::ivec2& size, int* posY, int &positionClimb) const;
-	bool canClimbDown(const glm::ivec2& pos, const glm::ivec2& size, int* posY, int& positionClimb) const;
+	bool canClimbUp(const glm::ivec2& pos, const glm::ivec2& size, float* posY, float &positionClimb) const;
+	bool canClimbDown(const glm::ivec2& pos, const glm::ivec2& size, float* posY, float & positionClimb) const;
 	
 
 	vector<Enemigo*> getEnemys();
